utilFunctions made redundant so fully removed
